backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.12.4
psycopg2-binary==2.8.5
sqlparse==0.4.2
tzdata==2022.1
pytz==2021.1
django-cors-headers==3.13.0